requirements.txt
firecrawl-py>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
